reorganized peer managing classes moving them to peers package
