+++ aeron-quick-base/src/main/java/org/jetc/aeron/quick/AeronQuickFactory.java
@@ -5,9 +5,10 @@ import io.aeron.CommonContext;
 import io.aeron.driver.MediaDriver;
 import io.aeron.exceptions.AeronException;
 import org.agrona.CloseHelper;
-import org.jetc.aeron.quick.sender.AeronQuickSenderBuilder;
-import org.jetc.aeron.quick.server.AeronQuickReceiverBuilder;
-import org.jetc.aeron.quick.server.precompile.ReceiverAdapterBase;
+import org.jetc.aeron.quick.peers.Adapters;
+import org.jetc.aeron.quick.peers.sender.AeronQuickSenderBuilder;
+import org.jetc.aeron.quick.peers.receiver.AeronQuickReceiverBuilder;
+import org.jetc.aeron.quick.peers.receiver.ReceiverAdapterBase;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import java.util.Optional;
