+++ aeron-quick-base/src/main/java/org/jetc/aeron/quick/peers/receiver/ReceiverAdapterBase.java
@@ -1,4 +1,4 @@
-package org.jetc.aeron.quick.server.precompile;
+package org.jetc.aeron.quick.peers.receiver;
 
 import org.jetc.aeron.quick.messaging.ReceiverBindingProvider;
