+++ aeron-quick-samples/src/main/java/org/jetc/aeron/quick/samples/first_basic_example/MyGeneralReceiver.java
@@ -2,7 +2,7 @@ package org.jetc.aeron.quick.samples.first_basic_example;
 
 import org.agrona.concurrent.SleepingMillisIdleStrategy;
 import org.jetc.aeron.quick.AeronQuickFactory;
-import org.jetc.aeron.quick.server.AeronQuickReceiverRunner;
+import org.jetc.aeron.quick.peers.receiver.AeronQuickReceiverRunner;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import java.time.Duration;
