+++ aeron-quick-samples/src/main/java/org/jetc/aeron/quick/samples/just_annotation_receiver/MyGeneralReceiver.java
@@ -4,7 +4,7 @@ import org.agrona.concurrent.SleepingMillisIdleStrategy;
 import org.jetc.aeron.quick.AeronQuickFactory;
 import org.jetc.aeron.quick.annotations.AeronQuickReceiver;
 import org.jetc.aeron.quick.annotations.QuickContractEndpoint;
-import org.jetc.aeron.quick.server.AeronQuickReceiverRunner;
+import org.jetc.aeron.quick.peers.receiver.AeronQuickReceiverRunner;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import java.time.Duration;
